backend/requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
async-generator==1.10
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.1
databases==0.5.5
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.0
fastapi==0.75.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==1.6
Mako==1.2.0
MarkupSafe==2.1.1
munch==2.5.0
numpy==1.22.3
orjson==3.6.8
outcome==1.1.0
packaging==21.3
pandas==1.4.2
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==3.5.3
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
slowapi==0.1.5
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a22
sqlmodel==0.0.6
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
trio==0.20.0
typing-extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
